+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
+## v2.1.0 2026-02-26
+### Added
+- `Number::$scale` property that count the decimal places in the number.
+- `Number::toFloat` method to cast the instance to `float` type.
+- `Number::toRadian()` method to convert the instance to a radian `Number`.
+- `Number::toDegree()` method to convert the instance to a degree `Number`.
+- `Number::rad()` method alias of `Number::toRadian()`.
+- `Number::deg()` method alias of `Number::toDegree()`.
+### Changed
+- API and README documentation.
+
 ## v2.0.1 2026-02-24
 ### Fixed
 - [#5](https://github.com/MarcoConsiglio/bcmath-extended/issues/5) The integer result of an expression is recognized as a decimal number